split project into packages
