+++ src/main/java/com/codeup/adlister/controllers/CounterServlet.java
@@ -1,14 +1,17 @@
+package com.codeup.adlister.controllers;
+
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "LogoutServlet", urlPatterns = "/logout")
-public class LogoutServlet extends HttpServlet {
+@WebServlet(name = "controllers.CounterServlet", urlPatterns = "/count")
+public class CounterServlet extends HttpServlet {
+    private int counter = 0;
+
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        request.getSession().removeAttribute("user");
-        request.getSession().invalidate();
-        response.sendRedirect("/login");
+        counter += 1;
+        response.getWriter().println("<h1>The count is " + counter + ".</h1>");
     }
 }
