+++ src/main/java/com/codeup/adlister/controllers/HelloWorldServlet.java
@@ -1,15 +1,14 @@
+package com.codeup.adlister.controllers;
+
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "CounterServlet", urlPatterns = "/count")
-public class CounterServlet extends HttpServlet {
-    private int counter = 0;
-
+@WebServlet(name = "controllers.HelloWorldServlet", urlPatterns = "/")
+public class HelloWorldServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        counter += 1;
-        response.getWriter().println("<h1>The count is " + counter + ".</h1>");
+        response.getWriter().println("<h1>Hello, World!</h1>");
     }
 }
