+++ src/main/java/com/codeup/adlister/controllers/LogoutServlet.java
@@ -1,12 +1,16 @@
+package com.codeup.adlister.controllers;
+
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "HelloWorldServlet", urlPatterns = "/")
-public class HelloWorldServlet extends HttpServlet {
+@WebServlet(name = "controllers.LogoutServlet", urlPatterns = "/logout")
+public class LogoutServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        response.getWriter().println("<h1>Hello, World!</h1>");
+        request.getSession().removeAttribute("user");
+        request.getSession().invalidate();
+        response.sendRedirect("/login");
     }
 }
