+++ src/main/java/com/codeup/adlister/dao/Ads.java
@@ -1,3 +1,7 @@
+package com.codeup.adlister.dao;
+
+import com.codeup.adlister.models.Ad;
+
 import java.util.List;
 
 public interface Ads {
