+++ src/main/java/com/codeup/adlister/dao/DaoFactory.java
@@ -1,3 +1,5 @@
+package com.codeup.adlister.dao;
+
 public class DaoFactory {
     private static Ads adsDao;
