+++ src/main/java/com/codeup/adlister/models/Ad.java
@@ -1,3 +1,5 @@
+package com.codeup.adlister.models;
+
 public class Ad {
     private long id;
     private long userId;
